AgentFennec/requirements.txt
requests
tiktoken
langchain==0.2.17
langchain-core
langchain-openai
langchain_community
langchain-text-splitters
networkx
PyYAML
whoosh
flask
flask-cors
openai
openpyxl
docx2txt
PyPDF2
python-pptx
setuptools
chikkarpy
jinja2
Nuitka